requirements.txt
streamlit==1.32.0
groq==0.4.1
google-generativeai==0.3.2